fix: comprehensive security hardening and code quality improvements (#29)
Address 34 issues from full-repository code review:

Backend (CRITICAL/HIGH):
- Sandbox code execution tool with module allowlist, builtin restrictions,
  and resource limits (CPU/memory/file size/file descriptors)
- Save AI replies on SSE stream interruption via try/finally
- Mask API keys in GET /api/settings response
- Add model_provider enum validation and content length limits
- Add rate limiting and field validation to unauthenticated log endpoint
- Move blocking bcrypt and extract_text calls to asyncio.to_thread
- Force OpenAI key for document embeddings regardless of LLM provider
- Rewrite logging middleware as pure ASGI (fixes streaming issues)
- Simplify Qdrant client to eager init with set-based collection tracking
- Disable unimplemented file tool from default enabled_tools
- Extend JWT expiration to 30 days

Frontend:
- Fix SSE streaming reactivity via array index access instead of stale ref
- Add line buffering and UTF-8 stream decoding for SSE chunks
- Support delta (incremental) SSE field with content fallback
- Add 401 response interceptor with auto-logout
- Add error handling and loading state to settings save
- Add i18n keys for saving/saveError/streamError across 6 languages

Infrastructure:
- Fix Redis healthcheck password leak (use REDISCLI_AUTH env var)
- Bind Grafana to localhost only
- Lock uv version in Dockerfile via multi-stage copy
- Change file_size_bytes to BigInteger with Alembic migration
- Add security documentation comments for privileged containers

Documentation:
- Correct CLAUDE.md: ports, route count, tool list, singleton description

+++ database/postgres/06_documents.sql
@@ -5,7 +5,7 @@ CREATE TABLE IF NOT EXISTS documents (
     user_id             UUID NOT NULL REFERENCES users(id) ON DELETE CASCADE,
     filename            VARCHAR(255) NOT NULL,
     file_type           VARCHAR(20) NOT NULL,
-    file_size_bytes     INTEGER NOT NULL,
+    file_size_bytes     BIGINT NOT NULL,
     chunk_count         INTEGER NOT NULL DEFAULT 0,
     qdrant_collection   VARCHAR(255) NOT NULL,
     minio_object_key    VARCHAR(500) NOT NULL,
